backend/pyproject.toml
[project]
name = "digital-mirror-backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.29",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "pydantic-settings>=2.2",
  "PyJWT[crypto]>=2.8",
  "httpx>=0.27",
  "boto3>=1.34",
  "redis[hiredis]>=5.0",
  "anthropic>=0.28",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "pytest-asyncio>=0.23",
  "anyio[trio]>=4.3",
  "ruff>=0.4",
  "mypy>=1.9",
  "boto3-stubs[s3]>=1.34",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
